cleanup rewrite exports
