+++ src/kirin/rewrite/chain.py
@@ -2,8 +2,7 @@ from typing import Iterable
 from dataclasses import dataclass
 
 from kirin.ir import IRNode
-from kirin.rewrite.abc import RewriteRule
-from kirin.rewrite.result import RewriteResult
+from kirin.rewrite.abc import RewriteRule, RewriteResult
 
 
 @dataclass
