+++ src/kirin/rewrite/walk.py
@@ -3,9 +3,8 @@ from dataclasses import field, dataclass
 
 from kirin.ir import Block, Region, Statement
 from kirin.worklist import WorkList
-from kirin.rewrite.abc import RewriteRule
+from kirin.rewrite.abc import RewriteRule, RewriteResult
 from kirin.ir.nodes.base import IRNode
-from kirin.rewrite.result import RewriteResult
 
 
 @dataclass
